Build Version v2.13.046 b

+++ appdriver/src/main/java/org/rmj/g3appdriver/lib/PetManager/model/LeaveApplication.java
@@ -78,7 +78,7 @@ public class LeaveApplication {
         Remarksxx = remarksxx;
     }
 
-    public int getNoOfHours(){
+    public double getNoOfHours(){
         noOfHours = noOfDaysx * 8;
         return noOfHours;
     }
